requirements.txt
streamsets
gitpython
